🎨 Refactor: Participant를  ParticipantList View로 분리
